+++ .config/webpack/script-tags.html
@@ -1,3 +1,4 @@
+<script type="text/javascript" src="{{ 'core.js' | asset_url }}"></script>
 <% for (var chunk in htmlWebpackPlugin.files.js) { %>
   <% var layoutEntry = htmlWebpackPlugin.files.js[chunk] %>
   <% var layoutEntry_name = layoutEntry.replace('../assets/', '').replace('.js', '').replace("layout-", '') %>
