+++ package.json
@@ -54,6 +54,7 @@
     "eslint-plugin-vue": "^7.9.0",
     "eslint-webpack-plugin": "^2.5.4",
     "file-loader": "^6.2.0",
+    "glob-all": "^3.2.1",
     "html-webpack-plugin": "^5.3.2",
     "mini-css-extract-plugin": "^1.5.0",
     "npm-run-all": "^4.1.5",
@@ -62,6 +63,7 @@
     "postcss-loader": "^5.2.0",
     "postcss-preset-env": "^6.7.0",
     "progress-webpack-plugin": "^1.0.12",
+    "purgecss-webpack-plugin": "^4.0.3",
     "sass": "^1.32.12",
     "sass-loader": "^11.0.1",
     "shopify-theme-lab-reloader": "^2.1.3",
