+++ shopify/layout/theme.liquid
@@ -15,19 +15,6 @@
 
     {{ content_for_header }}
 
-    <style>
-      #app {
-        min-width: 320px !important;
-      }
-      
-      #main {
-        margin-top: 75px;
-      }
-      
-      #shopify-section-footer {
-        position: relative;
-      }
-    </style>
     {% render 'style-tags' %}
   </head>
