+++ shopify/snippets/script-tags.liquid
@@ -1,3 +1,4 @@
+<script type="text/javascript" src="{{ 'core.js' | asset_url }}"></script>
