+++ shopify/snippets/style-tags.liquid
@@ -1,3 +1,4 @@
+<link type="text/css" href="{{ 'core.css' | | asset_url }}" rel="stylesheet">
