+++ src/layout/account.js
@@ -10,6 +10,14 @@ vueComponents.keys().forEach(key => {
   Vue.component(name, component);
 });
 
+/* Import Filters
+const filters = require.context("./vue/filters/", true, /\.(vue|js)$/);
+filters.keys().forEach(key => {
+  const filterMethod = filters(key);
+  const name = key.replace(/\.(\/|vue|js)/g, "").replace(/(\/|-|_|\s)\w/g, (match) => match.slice(1).toUpperCase());
+  Vue.filter(name, filterMethod);
+});*/
+
 new Vue({
   el: "#layout-account",
 });
