--- src/styles/bootstrap.scss
@@ -1,3 +1,0 @@
-
-@import 'node_modules/bootstrap/scss/bootstrap.scss';
-@import 'node_modules/bootstrap-vue/src/index.scss';
