+++ src/vue/core/main-nav/grouped-menu-item.vue
@@ -38,8 +38,13 @@
 </template>
 
 <script>
+//import { 	BCollapse, BIcon } from "bootstrap-vue";
 export default {
   name: "MainNavGroupedMenuItem",
+  components: {
+    //BCollapse,
+   // BIcon
+  },
   props: {
     name: {
       type: String,
